busqueda Local

+++ Programming/Java/SI/src/busqueda/Busqueda.java
@@ -7,15 +7,15 @@ public abstract class Busqueda<T> {
     protected Sucesor<T> sucesor;
     protected Goal<T> objetivo;
     protected int max_prof;
-    protected long max_obj_in_list;
-    protected long expanded_nodes;
+    protected int max_obj_in_list;
+    protected int expanded_nodes;
 
     public Busqueda(Sucesor<T> sucesor, Goal<T> objetivo, int max_prof ){
         this.sucesor = sucesor;
         this.objetivo = objetivo;
         this.max_prof = max_prof;
-        this.max_obj_in_list = 1L;
-        this.expanded_nodes = 0L;
+        this.max_obj_in_list = 1;
+        this.expanded_nodes = 0;
     }
 
     protected abstract ColeccionBusqueda<T> coleccion();
@@ -44,11 +44,11 @@ public abstract class Busqueda<T> {
         return null;
     }
 
-    public long getMax_obj_in_list() {
+    public int getMax_obj_in_list() {
         return max_obj_in_list;
     }
 
-    public long getExpanded_nodes() {
+    public int getExpanded_nodes() {
         return expanded_nodes;
     }
 }
